chore :: 기술 스택 버튼 수정 및 위치 변경
